requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
streamlit
PyPDF2